requirements.txt
asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
dj-database-url==3.0.1
Django==4.2.23
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.58.1
gunicorn==23.0.0
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
